MySQL initialize script change and script cleanup
