+++ tests/database/relational/resources/initialize_mysql.sql
@@ -15,7 +15,10 @@ CREATE TABLE IF NOT EXISTS task(
     PRIMARY KEY(taskId),
     FOREIGN KEY(projectId) REFERENCES project(projectId)
 );
-/*
+
+DELETE * FROM Project;
+DELETE * FROM Tasks;
+
 INSERT INTO Project(title, description) VALUES
 ("Organize Photos", "Organize old iPhone photos by year"),
 ("Read more", "Read book per month this year");
@@ -24,4 +27,3 @@ INSERT INTO Tasks(project_id, description) VALUES
 (1, "Organize 2019 photos"),
 (1, "Organize 2020 photos"),
 (2, "Read the Huntress");
-*/
